+++ frontend/src/component/map.tsx
@@ -1,106 +1,43 @@
 import React from "react";
-import { Wrapper, Status } from "@googlemaps/react-wrapper";
-import { createCustomEqual } from "fast-equals";
-import { isLatLngLiteral } from "@googlemaps/typescript-guards";
+import { Status, Wrapper } from "@googlemaps/react-wrapper";
+import config from "../config";
+import { useDeepCompareEffectForMaps } from "../hooks/map";
 
-const render = (status: Status) => {
-	return <h1>{status}</h1>;
-};
 
-const AppMap: React.VFC = () => {
-	// [START maps_react_map_component_app_state]
-	const [clicks, setClicks] = React.useState<google.maps.LatLng[]>([]);
-	const [zoom, setZoom] = React.useState(3); // initial zoom
-	const [center, setCenter] = React.useState<google.maps.LatLngLiteral>({
-		lat: 0,
-		lng: 0,
-	});
+interface MapProps extends google.maps.MapOptions {
+	style: { [key: string]: string };
+	onClick?: (e: google.maps.MapMouseEvent) => void;
+	onIdle?: (map: google.maps.Map) => void;
+}
 
-	const onClick = (e: google.maps.MapMouseEvent) => {
-		// avoid directly mutating state
-		setClicks([...clicks, e.latLng!]);
+
+const MapWrapper: React.VFC = () => {
+	const render = (status: Status) => {
+		switch (status) {
+			case Status.LOADING:
+				return <h1>Loading...</h1>;
+			case Status.FAILURE:
+				return <h1>Failed...</h1>;
+			default:
+				return <h1>Unknown...</h1>
+		}
 	};
+
+	const [zoom, setZoom] = React.useState(0);
+	const [center, setCenter] = React.useState<google.maps.LatLngLiteral>({
+		lat: -34,
+		lng: 151,
+	});
 
 	const onIdle = (m: google.maps.Map) => {
 		console.log("onIdle");
 		setZoom(m.getZoom()!);
 		setCenter(m.getCenter()!.toJSON());
 	};
-	// [END maps_react_map_component_app_state]
 
-	const form = (
-		<div
-			style={{
-				padding: "1rem",
-				flexBasis: "250px",
-				height: "100%",
-				overflow: "auto",
-			}}
-		>
-			<label htmlFor="zoom">Zoom</label>
-			<input
-				type="number"
-				id="zoom"
-				name="zoom"
-				value={zoom}
-				onChange={(event) => setZoom(Number(event.target.value))}
-			/>
-			<br />
-			<label htmlFor="lat">Latitude</label>
-			<input
-				type="number"
-				id="lat"
-				name="lat"
-				value={center.lat}
-				onChange={(event) =>
-					setCenter({ ...center, lat: Number(event.target.value) })
-				}
-			/>
-			<br />
-			<label htmlFor="lng">Longitude</label>
-			<input
-				type="number"
-				id="lng"
-				name="lng"
-				value={center.lng}
-				onChange={(event) =>
-					setCenter({ ...center, lng: Number(event.target.value) })
-				}
-			/>
-			<h3>{clicks.length === 0 ? "Click on map to add markers" : "Clicks"}</h3>
-			{clicks.map((latLng, i) => (
-				<pre key={i}>{JSON.stringify(latLng.toJSON(), null, 2)}</pre>
-			))}
-			<button onClick={() => setClicks([])}>Clear</button>
-		</div>
-	);
-
-	// [START maps_react_map_component_app_return]
-	return (
-		<div style={{ display: "flex", height: "100%" }}>
-			<Wrapper apiKey={process.env.REACT_APP_GOOGLE_MAPS_API_KEY} render={render}>
-				<Map
-					center={center}
-					onClick={onClick}
-					onIdle={onIdle}
-					zoom={zoom}
-					style={{ flexGrow: "1", height: "100%" }}
-				>
-					{clicks.map((latLng, i) => (
-						<Marker key={i} position={latLng} />
-					))}
-				</Map>
-			</Wrapper>
-			{/* Basic form for controlling center and zoom of map. */}
-			{form}
-		</div>
-	);
-	// [END maps_react_map_component_app_return]
-};
-interface MapProps extends google.maps.MapOptions {
-	style: { [key: string]: string };
-	onClick?: (e: google.maps.MapMouseEvent) => void;
-	onIdle?: (map: google.maps.Map) => void;
+	return <Wrapper apiKey={config.googleMapsApiKey} render={render}>
+		<Map style={{ width: "300px", height: "300px" }} zoom={zoom} center={center} onIdle={onIdle} />
+	</Wrapper>
 }
 
 const Map: React.FC<MapProps> = ({
@@ -110,7 +47,6 @@ const Map: React.FC<MapProps> = ({
 	style,
 	...options
 }) => {
-	// [START maps_react_map_component_add_map_hooks]
 	const ref = React.useRef<HTMLDivElement>(null);
 	const [map, setMap] = React.useState<google.maps.Map>();
 
@@ -119,9 +55,7 @@ const Map: React.FC<MapProps> = ({
 			setMap(new window.google.maps.Map(ref.current, {}));
 		}
 	}, [ref, map]);
-	// [END maps_react_map_component_add_map_hooks]
 
-	// [START maps_react_map_component_options_hook]
 	// because React does not do deep comparisons, a custom hook is used
 	// see discussion in https://github.com/googlemaps/js-samples/issues/946
 	useDeepCompareEffectForMaps(() => {
@@ -129,104 +63,8 @@ const Map: React.FC<MapProps> = ({
 			map.setOptions(options);
 		}
 	}, [map, options]);
-	// [END maps_react_map_component_options_hook]
 
-	// [START maps_react_map_component_event_hooks]
-	React.useEffect(() => {
-		if (map) {
-			["click", "idle"].forEach((eventName) =>
-				google.maps.event.clearListeners(map, eventName)
-			);
-
-			if (onClick) {
-				map.addListener("click", onClick);
-			}
-
-			if (onIdle) {
-				map.addListener("idle", () => onIdle(map));
-			}
-		}
-	}, [map, onClick, onIdle]);
-	// [END maps_react_map_component_event_hooks]
-
-	// [START maps_react_map_component_return]
-	return (
-		<>
-			<div ref={ref} style={style} />
-			{React.Children.map(children, (child) => {
-				if (React.isValidElement(child)) {
-					// set the map prop on the child component
-					return React.cloneElement(child, { map });
-				}
-			})}
-		</>
-	);
-	// [END maps_react_map_component_return]
+	return <div ref={ref} style={style} />
 };
 
-// [START maps_react_map_marker_component]
-const Marker: React.FC<google.maps.MarkerOptions> = (options) => {
-	const [marker, setMarker] = React.useState<google.maps.Marker>();
-
-	React.useEffect(() => {
-		if (!marker) {
-			setMarker(new google.maps.Marker());
-		}
-
-		// remove marker from map on unmount
-		return () => {
-			if (marker) {
-				marker.setMap(null);
-			}
-		};
-	}, [marker]);
-
-	React.useEffect(() => {
-		if (marker) {
-			marker.setOptions(options);
-		}
-	}, [marker, options]);
-
-	return null;
-};
-// [END maps_react_map_marker_component]
-
-const deepCompareEqualsForMaps = createCustomEqual(
-	(deepEqual: any) => (a: any, b: any) => {
-		if (
-			isLatLngLiteral(a) ||
-			a instanceof google.maps.LatLng ||
-			isLatLngLiteral(b) ||
-			b instanceof google.maps.LatLng
-		) {
-			return new google.maps.LatLng(a).equals(new google.maps.LatLng(b));
-		}
-
-		// TODO extend to other types
-
-		// use fast-equals for other objects
-		return deepEqual(a, b);
-	}
-);
-
-function useDeepCompareMemoize(value: any) {
-	const ref = React.useRef();
-
-	if (!deepCompareEqualsForMaps(value, ref.current)) {
-		ref.current = value;
-	}
-
-	return ref.current;
-}
-
-function useDeepCompareEffectForMaps(
-	callback: React.EffectCallback,
-	dependencies: any[]
-) {
-	React.useEffect(callback, dependencies.map(useDeepCompareMemoize));
-}
-
-// [END maps_react_map]
-let PRESERVE_COMMENT_ABOVE; // force tsc to maintain the comment above eslint-disable-line
-
-export default AppMap;
+export default MapWrapper;
