+++ frontend/src/config.ts
@@ -1,0 +1,5 @@
+const config = {
+    googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY || "unconfigured"
+}
+
+export default config;
