+++ frontend/tsconfig.json
@@ -7,7 +7,8 @@
       "esnext"
     ],
     "types": [
-      "googlemaps"
+      "googlemaps",
+      "node"
     ],
     "allowJs": true,
     "skipLibCheck": true,
